+++ NEWS.md
@@ -1,5 +1,6 @@
 # forestTIME-builder (development version)
 
+- `interpolate_data()` no longer produces negative values for `HT`, `DIA`, or `ACTUALHT`.  Instead, trees that get extrapolated to have values ≤ 0 are assumed to be fallen and dead (`STATUSCD` 2 and `STANDING_DEAD_CD` 0).  These fallen dead trees then have their measurements set to `NA` by `adjust_mortality()`.
 - `prep_data()` now converts `PLT_CN` from numeric to character for better readability in the output.
 - Empty plots are no longer dropped silently by `prep_data()` and should be handled correctly by the rest of the workflow through `interpolate_data()`.
 - `expand_data()` now adds a column, `interpolated`, that marks whether an observation was interpolated (`TRUE`) or in the original data (`FALSE`).